Add dashboard to the ps landing page
